requirements.txt
pytest
tensorflow>=2.0.0
torch>=1.9.0
transformers>=4.21.0
Pillow>=8.0.0
numpy>=1.19.0
opencv-python>=4.5.0
scikit-learn>=1.0.0 